+++ include/lss/game/creature.hpp
@@ -63,7 +63,21 @@ public:
   std::optional<std::tuple<std::shared_ptr<Slot>, int, int, int>>
       getSecondaryDmg(std::shared_ptr<Slot>);
   bool hasTrait(Trait t) {
-    return std::find(traits.begin(), traits.end(), t) != traits.end();
+    auto allTraits = traits;
+
+  for (auto s : equipment->slots) {
+    if (s->item == nullptr ||
+        std::find(s->acceptTypes.begin(), s->acceptTypes.end(),
+                  s->item->type.wearableType) == s->acceptTypes.end()) {
+      continue;
+    }
+    for (auto e : s->item->effects) {
+      if(auto et = std::dynamic_pointer_cast<TraitEffect>(e)){
+        allTraits.push_back(et->trait);
+      }
+    }
+  }
+    return std::find(allTraits.begin(), allTraits.end(), t) != allTraits.end();
   }
   bool hasLight();
   std::shared_ptr<Damage> updateDamage(std::shared_ptr<Damage>, int m, int d,